Remove unused extension of Railtie to import rake tasks
Summary:
Previously, we were importing 'properties.rake' task for end users.
This file was renamed to 'hubspot.rake', which led to users encountering
the LoadError "cannot load such file" since the import referenced a file
that no longer existed.

Rather than add the import back to reference the correct file,
'hubspot.rake', this change removes the railtie extension as we discuss
the future of 'hubspot.rake' and if we wish to continue supporting it or
remove it.