Establish the use of a changelog
Summary:
Why keep a changelog?
Changelogs help users and contributors see what notable changes have
been made in each version of the project.

Why is this changelog named History.md?
I opted to use the name History over changelog as the purpose of this
file is to be a user-facing list of changes rather than a list of
changes that do not impact the user (like test improvements, etc)

Changelog conventions:
- Each version should have an entry, or entries, in the changelog
- The latest version is at the top of the file
- The release date of each version is displayed